+++ flutter/lib/src/navigation/sentry_navigator_observer.dart
@@ -35,16 +35,19 @@ const _navigationKey = 'navigation';
 ///   - [RouteObserver](https://api.flutter.dev/flutter/widgets/RouteObserver-class.html)
 ///   - [Navigating with arguments](https://flutter.dev/docs/cookbook/navigation/navigate-with-arguments)
 class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
-  SentryNavigatorObserver({Hub? hub, this.setTransaction = true})
+  SentryNavigatorObserver({Hub? hub, this.setRouteNameAsTransaction = true})
       : _hub = hub ?? HubAdapter();
 
   final Hub _hub;
-  final bool setTransaction;
+
+  /// Enabling this option overrides the current [Scope.transaction].
+  /// So be careful when this is used together with performance tracing.
+  final bool setRouteNameAsTransaction;
 
   @override
   void didPush(Route<dynamic> route, Route<dynamic>? previousRoute) {
     super.didPush(route, previousRoute);
-    setCurrentRoute(route.settings.name);
+    _setCurrentRoute(route.settings.name);
     _addBreadcrumb(
       type: 'didPush',
       from: previousRoute?.settings,
@@ -55,7 +58,7 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
   @override
   void didReplace({Route<dynamic>? newRoute, Route<dynamic>? oldRoute}) {
     super.didReplace(newRoute: newRoute, oldRoute: oldRoute);
-    setCurrentRoute(newRoute?.settings.name);
+    _setCurrentRoute(newRoute?.settings.name);
     _addBreadcrumb(
       type: 'didReplace',
       from: oldRoute?.settings,
@@ -66,7 +69,7 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
   @override
   void didPop(Route<dynamic> route, Route<dynamic>? previousRoute) {
     super.didPop(route, previousRoute);
-    setCurrentRoute(previousRoute?.settings.name);
+    _setCurrentRoute(previousRoute?.settings.name);
     _addBreadcrumb(
       type: 'didPop',
       from: route.settings,
@@ -86,8 +89,11 @@ class SentryNavigatorObserver extends RouteObserver<PageRoute<dynamic>> {
     ));
   }
 
-  void setCurrentRoute(String? name) {
-    if (setTransaction) {
+  void _setCurrentRoute(String? name) {
+    if (name == null) {
+      return;
+    }
+    if (setRouteNameAsTransaction) {
       _hub.configureScope((scope) {
         scope.transaction = name;
       });